+++ packages/shared/src/contexts/payment/Paddle.tsx
@@ -1,18 +1,12 @@
-import type { ReactElement, ReactNode } from 'react';
+import type { ReactElement } from 'react';
 import React, {
   useCallback,
-  useContext,
   useEffect,
   useMemo,
   useRef,
   useState,
 } from 'react';
-import type {
-  Environments,
-  Paddle,
-  PaddleEventData,
-  TimePeriod,
-} from '@paddle/paddle-js';
+import type { Environments, Paddle, PaddleEventData } from '@paddle/paddle-js';
 import {
   CheckoutEventNames,
   getPaddleInstance,
@@ -24,177 +18,19 @@ import { useAuthContext } from '../AuthContext';
 import { plusSuccessUrl } from '../../lib/constants';
 import { LogEvent } from '../../lib/log';
 import { usePlusSubscription } from '../../hooks';
-import { feature, featureIAPProducts } from '../../lib/featureManagement';
+import { feature } from '../../lib/featureManagement';
 import { PlusPriceType, PlusPriceTypeAppsId } from '../../lib/featureValues';
 import { getPrice } from '../../lib';
 import { useFeature } from '../../components/GrowthBookProvider';
-import {
-  checkIsExtension,
-  isIOSNative,
-  isNullOrUndefined,
-  promisifyEventListener,
-} from '../../lib/func';
+import { checkIsExtension } from '../../lib/func';
 import { usePixelsContext } from '../PixelsContext';
-import { sendMessage, WebKitMessageHandlers } from '../../lib/ios';
-
-export type ProductOption = {
-  label: string;
-  value: string;
-  price: {
-    amount: number;
-    formatted: string;
-    monthlyAmount: number;
-    monthlyFormatted: string;
-  };
-  currencyCode?: string;
-  currencySymbol?: string;
-  extraLabel: string;
-  appsId: PlusPriceTypeAppsId;
-  duration: PlusPriceType;
-  durationLabel: 'month' | 'year';
-  trialPeriod: TimePeriod | null;
-};
-
-interface OpenCheckoutProps {
-  priceId: string;
-  giftToUserId?: string;
-}
-
-export type OpenCheckoutFn = (props: OpenCheckoutProps) => void;
-
-export interface PaymentContextData {
-  openCheckout?: OpenCheckoutFn;
-  paddle?: Paddle | undefined;
-  productOptions?: ProductOption[];
-  earlyAdopterPlanId?: string | null;
-  isPlusAvailable: boolean;
-  giftOneYear?: ProductOption;
-  isPricesPending: boolean;
-  isFreeTrialExperiment: boolean;
-}
-
-const PaymentContext = React.createContext<PaymentContextData>(undefined);
-export default PaymentContext;
-
-export type PaymentContextProviderProps = {
-  children?: ReactNode;
-};
-
-export type IAPProduct = {
-  attributes: {
-    description: {
-      standard: string;
-    };
-    icuLocale: string;
-    isFamilyShareable: number;
-    kind: string;
-    name: string;
-    offerName: string;
-    offers: {
-      currencyCode: string;
-      discounts: unknown[]; // Specify a more detailed type if known
-      price: string;
-      priceFormatted: string;
-      recurringSubscriptionPeriod: string;
-    }[];
-    subscriptionFamilyId: string;
-    subscriptionFamilyName: string;
-    subscriptionFamilyRank: number;
-  };
-  href: string;
-  id: string;
-  type: string;
-};
-
-export const StoreKitSubProvider = ({
-  children,
-}: PaymentContextProviderProps): ReactElement => {
-  const { user } = useAuthContext();
-  const productIds = useFeature(featureIAPProducts);
-  const productList = useMemo(() => Object.keys(productIds), [productIds]);
-
-  const { data: productOptions } = useQuery({
-    queryKey: ['iap-products'],
-    enabled: !!productIds,
-    queryFn: async () => {
-      const products = promisifyEventListener(
-        'iap-products-result',
-        (event) => {
-          const productsRaw = !isNullOrUndefined(event?.detail)
-            ? JSON.parse(event.detail)
-            : [];
-
-          return productsRaw
-            ?.map((product: IAPProduct): ProductOption => {
-              const duration = productIds[
-                product.attributes.offerName
-              ] as PlusPriceType;
-
-              return {
-                label: product.attributes.name,
-                value: product.attributes.offerName,
-                price: {
-                  amount: parseFloat(product.attributes.offers[0].price),
-                  formatted: product.attributes.offers[0].priceFormatted,
-                  monthlyAmount: parseFloat(product.attributes.offers[0].price),
-                  monthlyFormatted: product.attributes.offers[0].priceFormatted,
-                },
-                extraLabel: product.attributes?.description?.standard,
-                appsId:
-                  product.attributes.offerName === 'annualSpecial'
-                    ? PlusPriceTypeAppsId.EarlyAdopter
-                    : PlusPriceTypeAppsId.Default,
-                duration,
-                durationLabel:
-                  duration === PlusPriceType.Yearly ? 'year' : 'month',
-                trialPeriod: null,
-              };
-            })
-            .sort((a: { value: string }, b: { value: string }) => {
-              // Make sure that the products are sorted in the same order as the product list
-              // because the native code does not guarantee the order of the products
-              const aIndex = productList.indexOf(a.value);
-              const bIndex = productList.indexOf(b.value);
-              return aIndex - bIndex;
-            });
-        },
-      );
-
-      sendMessage(WebKitMessageHandlers.IAPProductList, productList);
-
-      return products;
-    },
-  });
-
-  const openCheckout = useCallback(
-    ({ priceId }: OpenCheckoutProps) => {
-      sendMessage(WebKitMessageHandlers.IAPSubscriptionRequest, {
-        productId: priceId,
-        appAccountToken: user?.subscriptionFlags?.appAccountToken,
-      });
-    },
-    [user?.subscriptionFlags?.appAccountToken],
-  );
-
-  const contextData = useMemo<PaymentContextData>(
-    () => ({
-      openCheckout,
-      productOptions,
-      earlyAdopterPlanId: null,
-      isPlusAvailable: false,
-      giftOneYear: undefined,
-      isPricesPending: false,
-      isFreeTrialExperiment: false,
-    }),
-    [openCheckout, productOptions],
-  );
-
-  return (
-    <PaymentContext.Provider value={contextData}>
-      {children}
-    </PaymentContext.Provider>
-  );
-};
+import type {
+  PaymentContextProviderProps,
+  OpenCheckoutProps,
+  PaymentContextData,
+  ProductOption,
+} from './context';
+import { PaymentContext } from './context';
 
 export const PaddleSubProvider = ({
   children,
@@ -438,16 +274,3 @@ export const PaddleSubProvider = ({
     </PaymentContext.Provider>
   );
 };
-
-export const PaymentContextProvider = ({
-  children,
-}: PaymentContextProviderProps): ReactElement => {
-  if (isIOSNative()) {
-    return <StoreKitSubProvider>{children}</StoreKitSubProvider>;
-  }
-
-  return <PaddleSubProvider>{children}</PaddleSubProvider>;
-};
-
-export const usePaymentContext = (): PaymentContextData =>
-  useContext(PaymentContext);
